Added platformer character controller
It's pretty bad, but goes to show how difficult writing something like this can be.

Needs some changes still, movementForce shouldn't be in Input probably, input should just pass "go right" or "go left" as an float, and we map that to the force. That way joysticks can be analog, but the controller controls the speed.